extract cell adapter as preparation for multi table
