+++ src/styles/_vars.scss
@@ -1,4 +1,4 @@
-$engine_selector: ".lineup-engine" !default;
+$engine_selector: '.lineup-engine' !default;
 $engine_assets: "./assets" !default;
 $engine_loading_image: url('#{$engine_assets}/loading.svg') !default;
 $engine_loading_static_image: url('#{$engine_assets}/loading_s.svg') !default;
